+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -133,6 +133,16 @@ New checks
   Checks that all implicit and explicit inline functions in header files are
   tagged with the ``LIBC_INLINE`` macro.
 
+- New :doc:`portability-non-portable-integer-constant
+  <clang-tidy/checks/portability-non-portable-integer-constant>` check.
+
+  Finds masks that are being used in a non-portable manner.
+
+- New :doc:`portability-non-portable-integer-constant
+  <clang-tidy/checks/portability/non-portable-integer-constant>` check.
+
+  Finds integer literals that are being used in a non-portable manner.
+
 New check aliases
 ^^^^^^^^^^^^^^^^^